move remaining commands to command package
